Refactor block.py and file_writer.py

+++ recs/audio/silence.py
@@ -2,7 +2,7 @@ import dataclasses as dc
 import typing as t
 from functools import cached_property
 
-T = t.TypeVar('T', int, float)
+T = t.TypeVar('T', float, int)
 
 
 @dc.dataclass(frozen=True)
